* fixed the timescale bug after savegame loading * buildings are now also stored in savegame * two new function to verifiy the loading and saving process of a base   buildinglist - prints the building matrices of all founded bases and other stuff   baselist     - prints the base matrix with building ids and other stuff * renamed addWorkers to assignedWorkers in building_t * fixed warnings in common.c (SZ_Buffer related)

+++ src/qcommon/common.c
@@ -475,22 +475,20 @@ void MSG_WriteFormat( sizebuf_t *sb, char *format, ... )
 			MSG_WriteLong( sb, va_arg( ap, int ) );
 			break;
 		case 'f':
-// 			MSG_WriteFloat( sb, va_arg( ap, float ) );
+			// NOTE: float is promoted to double through ...
 			MSG_WriteFloat( sb, va_arg( ap, double ) );
 			break;
 		case 'p':
-// 			MSG_WritePos( sb, va_arg( ap, float* ) );
-			MSG_WritePos( sb, va_arg( ap, double* ) );
+ 			MSG_WritePos( sb, va_arg( ap, float* ) );
 			break;
 		case 'g':
 			MSG_WriteGPos( sb, va_arg( ap, byte* ) );
 			break;
 		case 'd':
-// 			MSG_WriteDir( sb, va_arg( ap, float* ) );
-			MSG_WriteDir( sb, va_arg( ap, double* ) );
+			MSG_WriteDir( sb, va_arg( ap, float* ) );
 			break;
 		case 'a':
-// 			MSG_WriteAngle( sb, va_arg( ap, float ) );
+			// NOTE: float is promoted to double through ...
 			MSG_WriteAngle( sb, va_arg( ap, double ) );
 			break;
 		case '!':
